+++ addons/payment/tests/__init__.py
@@ -2,7 +2,6 @@
 
 from . import common
 from . import http_common
-from . import multicompany_common
 from . import test_flows
 from . import test_multicompany_flows
 from . import test_payments
